Skilt Services fra Clients og lagt sammen med Mappers under services. Tatt inn ArchUnit-tester for å verifisere at vi ikke bruker noe i clients i domenet eller i regler. Fikset de stedene vi gjorde det
