+++ src/main/kotlin/no/nav/medlemskap/clients/Services.kt
@@ -1,20 +1,14 @@
 package no.nav.medlemskap.clients
 
 import no.nav.medlemskap.clients.aareg.AaRegClient
-import no.nav.medlemskap.clients.aareg.AaRegService
 import no.nav.medlemskap.clients.ereg.EregClient
 import no.nav.medlemskap.clients.medl.MedlClient
-import no.nav.medlemskap.clients.medl.MedlService
 import no.nav.medlemskap.clients.oppgave.OppgaveClient
-import no.nav.medlemskap.clients.oppgave.OppgaveService
 import no.nav.medlemskap.clients.pdl.PdlClient
-import no.nav.medlemskap.clients.pdl.PdlService
 import no.nav.medlemskap.clients.saf.SafClient
-import no.nav.medlemskap.clients.saf.SafService
 import no.nav.medlemskap.clients.sts.StsRestClient
 import no.nav.medlemskap.clients.sts.stsClient
 import no.nav.medlemskap.clients.tpsws.PersonClient
-import no.nav.medlemskap.clients.tpsws.PersonService
 import no.nav.medlemskap.common.callIdGenerator
 import no.nav.medlemskap.common.cioHttpClient
 import no.nav.medlemskap.common.healthcheck.HealthReporter
@@ -23,6 +17,12 @@ import no.nav.medlemskap.common.healthcheck.HttpResponseHealthCheck
 import no.nav.medlemskap.common.healthcheck.TryCatchHealthCheck
 import no.nav.medlemskap.config.Configuration
 import no.nav.medlemskap.config.retryRegistry
+import no.nav.medlemskap.services.aareg.AaRegService
+import no.nav.medlemskap.services.medl.MedlService
+import no.nav.medlemskap.services.oppgave.OppgaveService
+import no.nav.medlemskap.services.pdl.PdlService
+import no.nav.medlemskap.services.saf.SafService
+import no.nav.medlemskap.services.tps.PersonService
 
 class Services(val configuration: Configuration) {
