+++ src/main/kotlin/no/nav/medlemskap/clients/medl/MedlClient.kt
@@ -68,8 +68,3 @@ class MedlClient(
 
 }
 
-class MedlService(private val medlClient: MedlClient) {
-
-    suspend fun hentMedlemskapsunntak(ident: String, callId: String, fraOgMed: LocalDate? = null, tilOgMed: LocalDate? = null) =
-            mapMedlemskapResultat(medlClient.hentMedlemskapsunntak(ident, callId, fraOgMed, tilOgMed))
-}
