+++ src/main/kotlin/no/nav/medlemskap/clients/oppgave/OppgaveClient.kt
@@ -52,9 +52,3 @@ class OppgaveClient(
 
 }
 
-class OppgaveService(private val oppgaveClient: OppgaveClient) {
-
-    suspend fun hentOppgaver(aktorIder: List<String>, callId: String) =
-            mapOppgaveResultat(oppgaveClient.hentOppgaver(aktorIder, callId).oppgaver)
-
-}
