+++ src/main/kotlin/no/nav/medlemskap/domene/Ansatte.kt
@@ -1,0 +1,6 @@
+package no.nav.medlemskap.domene
+
+data class Ansatte(
+        val antall: Int?,
+        val bruksperiode: Bruksperiode?,
+        val gyldighetsperiode: Gyldighetsperiode?)
