+++ src/main/kotlin/no/nav/medlemskap/domene/OpplysningspliktigArbeidsgiverType.kt
@@ -1,0 +1,5 @@
+package no.nav.medlemskap.domene
+
+enum class OpplysningspliktigArbeidsgiverType {
+    Organisasjon, Person
+}
