+++ src/main/kotlin/no/nav/medlemskap/services/aareg/AaregMapper.kt
@@ -1,8 +1,11 @@
-package no.nav.medlemskap.clients.aareg
+package no.nav.medlemskap.services.aareg
 
 import mu.KotlinLogging
+import no.nav.medlemskap.clients.aareg.AaRegArbeidsavtale
+import no.nav.medlemskap.clients.aareg.AaRegArbeidsforhold
+import no.nav.medlemskap.clients.aareg.AaRegOpplysningspliktigArbeidsgiverType
+import no.nav.medlemskap.clients.aareg.AaRegUtenlandsopphold
 import no.nav.medlemskap.domene.*
-
 
 private val logger = KotlinLogging.logger { }
 
@@ -12,12 +15,18 @@ fun mapAaregResultat(arbeidsforhold: List<AaRegArbeidsforhold>, dataOmArbeidsgiv
                 periode = mapPeriodeTilArbeidsforhold(it),
                 utenlandsopphold = mapUtenLandsopphold(it),
                 arbeidsfolholdstype = mapArbeidsForholdType(it),
-                arbeidsgivertype = it.arbeidsgiver.type,
+                arbeidsgivertype = mapArbeidsgiverType(it.arbeidsgiver.type),
                 arbeidsgiver = mapArbeidsgiver(it, dataOmArbeidsgiver),
                 arbeidsavtaler = mapArbeidsAvtaler(it)
         )
     }
 }
+
+fun mapArbeidsgiverType(type: AaRegOpplysningspliktigArbeidsgiverType): OpplysningspliktigArbeidsgiverType =
+        when (type) {
+            AaRegOpplysningspliktigArbeidsgiverType.Person -> OpplysningspliktigArbeidsgiverType.Person
+            AaRegOpplysningspliktigArbeidsgiverType.Organisasjon -> OpplysningspliktigArbeidsgiverType.Organisasjon
+        }
 
 fun mapPeriodeTilArbeidsforhold(arbeidsforhold: AaRegArbeidsforhold): Periode {
     return Periode(
@@ -71,9 +80,24 @@ fun mapArbeidsgiver(arbeidsforhold: AaRegArbeidsforhold, dataOmArbeidsgiver: Mut
     return Arbeidsgiver(
             type = enhetstype,
             identifikator = orgnummer,
-            ansatte = ansatte,
+            ansatte = mapAnsatte(ansatte),
             konkursStatus = konkursStatus)
 }
+
+fun mapAnsatte(ansatte: List<no.nav.medlemskap.clients.ereg.Ansatte>?): List<Ansatte>? =
+        ansatte?.map {
+            Ansatte(
+                    antall = it.antall,
+                    bruksperiode = mapBruksperiode(it.bruksperiode),
+                    gyldighetsperiode = mapGyldighetsperiode(it.gyldighetsperiode)
+            )
+        }
+
+fun mapGyldighetsperiode(gyldighetsperiode: no.nav.medlemskap.clients.ereg.Gyldighetsperiode?): Gyldighetsperiode? =
+        gyldighetsperiode?.let { Gyldighetsperiode(fom = it.fom, tom = it.tom) }
+
+fun mapBruksperiode(bruksperiode: no.nav.medlemskap.clients.ereg.Bruksperiode?): Bruksperiode? =
+        bruksperiode?.let { Bruksperiode(fom = it.fom, tom = it.tom) }
 
 fun mapPeriodeTilArbeidsavtale(arbeidsavtale: AaRegArbeidsavtale): Periode {
     return Periode(
