+++ src/main/kotlin/no/nav/medlemskap/services/saf/SafMapper.kt
@@ -1,4 +1,4 @@
-package no.nav.medlemskap.clients.saf
+package no.nav.medlemskap.services.saf
 
 import no.nav.medlemskap.client.generated.Dokumenter
 import no.nav.medlemskap.domene.Dokument
