+++ src/main/kotlin/no/nav/medlemskap/services/saf/SafService.kt
@@ -1,0 +1,8 @@
+package no.nav.medlemskap.services.saf
+
+import no.nav.medlemskap.clients.saf.SafClient
+
+class SafService(private val safClient: SafClient) {
+    suspend fun hentJournaldata(fnr: String, callId: String) =
+            mapDokumentoversiktBrukerResponse(safClient.hentJournaldata(fnr, callId))
+}
