+++ src/test/kotlin/no/nav/medlemskap/clients/pdl/PdlSivilstandMapperTest.kt
@@ -1,6 +1,7 @@
 package no.nav.medlemskap.clients.pdl
 
 import no.nav.medlemskap.domene.Sivilstandstype
+import no.nav.medlemskap.services.pdl.PdlSivilstandMapper
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNull
 import org.junit.jupiter.api.Test
